refactor: Refactor directory structure of components
The components previously nested under 'components' directories have been moved up a level. They now reside directly within their respective route directories. This makes their locations in the codebase more intuitive and reduces unnecessary depth in the directory structure. All relevant import paths and usage locations have also been updated accordingly to reflect these changes.
