--- src/routes/about-me/components/index.tsx
@@ -1,4 +1,0 @@
-/** @package */
-export { Activitybar } from "./activitybar";
-/** @package */
-export { Sidebar } from "./sidebar";
