--- src/routes/hello/components/index.tsx
@@ -1,1 +1,0 @@
-export { Game } from "./game";
