+++ src/routes/hello/game/snake-placeholder.tsx
